+++ app/controllers/api/users_controller.rb
@@ -1,13 +1,17 @@
 module Api
   class UsersController < ApplicationController
+    skip_before_action :authenticate, only: [:create]
+
     def index
       @users = User.all
+      authorize @users
 
       render_with_root(request.headers['X_API_SERIALIZER_ROOT'])
     end
 
     def show
       @user = User.find(params[:id])
+      authorize @user
 
       render_with_serializer(request.headers['X_API_SERIALIZER'])
     end
@@ -22,17 +26,25 @@ module Api
       end
     end
 
-    def update
-      @user = User.find(params[:id])
-      if @user.update(user_params)
-        render json: UserSerializer.render(@user, root: :user)
+    def update # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
+      if (params[:id].to_i != current_user.id || params[:user][:role].present?) &&
+         current_user.role != 'admin'
+        render json: { error: 'Authorization denied' }, status: :forbidden
       else
-        render json: { errors: @user.errors.as_json }, status: :bad_request
+        @user = User.find(params[:id])
+        authorize @user
+
+        if @user.update(user_params)
+          render json: UserSerializer.render(@user, root: :user)
+        else
+          render json: { errors: @user.errors.as_json }, status: :bad_request
+        end
       end
     end
 
     def destroy
       @user = User.find(params[:id])
+      authorize @user
 
       if @user.destroy
         render json: { messages: ['User has been deleted.'] }, status: :no_content
@@ -44,7 +56,7 @@ module Api
     private
 
     def user_params
-      params.require(:user).permit(:first_name, :last_name, :email)
+      params.require(:user).permit(:first_name, :last_name, :email, :password, :role)
     end
   end
 end
